DevFest2025/3-imageDescriberJob/requirements.txt
google-cloud-storage==3.4.1
google-cloud-aiplatform==1.124.0
